Refactor by fragments
