+++ app/src/main/java/com/example/hueray/MainActivity.kt
@@ -3,7 +3,6 @@ package com.example.hueray
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.hueray.happyscore.database.HappyScoreDatabase
-import com.example.hueray.happyscore.viewmodel.HappyScoreViewModel
 
 
 class MainActivity : AppCompatActivity() {
