+++ app/src/main/java/com/example/hueray/happyscore/repository/HueRayRepository.kt
@@ -7,7 +7,7 @@ import com.example.hueray.happyscore.database.HappyScoreDao
 
 // Declares the DAO as a private property in the constructor. Pass in the DAO
 // instead of the whole database, because you only need access to the DAO
-class HappyScoreRepository(private val HappyScoreDao: HappyScoreDao) {
+class HueRayRepository(private val HappyScoreDao: HappyScoreDao) {
 
     // Room executes all queries on a separate thread.
     // Observed LiveData will notify the observer when the data has changed.
