+++ app/src/main/res/layout/fragment_rate_your_happiness.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".rate_your_happiness">
+    tools:context=".fragments.happy.HappyRate">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
