+++ app/src/main/res/layout/fragment_thanks_for_rating.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".thanks_for_rating">
+    tools:context=".fragments.happy.HappyScores">
 
     <!-- TODO: Update blank fragment layout -->
